+++ spec/knossos/algorithm/wilsons_spec.rb
@@ -1,0 +1,5 @@
+require 'knossos/algorithm/shared_examples_for_algorithm'
+
+RSpec.describe Algorithm::Wilsons do
+  it_behaves_like 'algorithm'
+end